dom/html/test/test_bug486741.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=486741
-->
<head>
  <title>Test for Bug 486741</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=486741">Mozilla Bug 486741</a>
<p id="display"><iframe id="f"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 486741 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var d = $("f").contentDocument;
  var root = d.documentElement;
  is(root.tagName, "HTML", "Unexpected root");

  d.open();
  isnot(d.documentElement, root, "Shouldn't have the old root element");

  d.write("Test");
  d.close();

  isnot(d.documentElement, root, "Still shouldn't have the old root element");
  is(d.documentElement.tagName, "HTML", "Unexpected new root after write");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
